Fix migrations when edition is active
After 48c8db6407e995c9e374850b9103c6b897c05930 it is possible to initially run migrations and create new migrations.

However, when an edition is active, creating new migrations will still fail because the edition class can not be serialized as the default value for other classes.

Marking the Edition class as [deconstructible](https://docs.djangoproject.com/en/3.1/topics/migrations/#adding-a-deconstruct-method) is an option, but not ideal, as this will still generate a migration for every new edition (even if this is not really changing anything).

I still think this is preferable over not being able to migrate, so I
propose this PR.